semi-optional parameters in submitBlock etc are now long

+++ src/linearoperator.hpp
@@ -45,7 +45,7 @@ public:
 	 *              Note that this must not be the block-column index for any matrix type.
 	 */
 	virtual void submitBlock(const index starti, const index startj, 
-		const scalar *const buffer, const size_t param1, const size_t param2) = 0;
+		const scalar *const buffer, const long param1, const long param2) = 0;
 	
 	/// Supposed to add to a contiguous block of the matrix in a thread-safe manner
 	/**
@@ -55,7 +55,7 @@ public:
 	 *              Note that this must not be the block-column index for any matrix type.
 	 */
 	virtual void updateBlock(const index starti, const index startj, 
-		const scalar *const buffer, const size_t param1, const size_t param2) = 0;
+		const scalar *const buffer, const long param1, const long param2) = 0;
 	
 	/// Supposed to update diagonal entries for point matrices 
 	/// and diagonal blocks for block matrices
